+++ ui-dist/assets/app.settings-BumMDSDR.js
@@ -1,1 +1,0 @@
-import{O as o}from"./index-C6ReVsUB.js";const n=o;export{n as component};