+++ ui-dist/assets/app.settings-DwaQibEv.js
@@ -1,0 +1,1 @@
+import{O as o}from"./index-D-eG31Tu.js";const n=o;export{n as component};