+++ ui-dist/assets/copy-JcLmFzcb.js
@@ -1,4 +1,4 @@
-import{c}from"./createLucideIcon-Cnq3OevU.js";/**
+import{c}from"./createLucideIcon-DMDLbIDM.js";/**
  * @license lucide-react v0.575.0 - ISC
  *
  * This source code is licensed under the ISC license.