+++ ui-dist/assets/preferences-DuMif6oN.js
@@ -1,4 +1,4 @@
-import{c as r}from"./createLucideIcon-Cnq3OevU.js";import{A as n,B as c}from"./api-Db5FdPZQ.js";/**
+import{c as r}from"./createLucideIcon-DMDLbIDM.js";import{A as n,B as c}from"./api-DnYlYPLj.js";/**
  * @license lucide-react v0.575.0 - ISC
  *
  * This source code is licensed under the ISC license.